refactor(observability): 分离体验复核报告
